Properly handle recycling of RecomposeHighlighterModifier (#510)
* Properly handle recycling of RecomposeHighlighterModifier

After migrating to be Node based, it's possible the Modifier node can be reused. For example, when scrolling in a LazyColumn the nodes of items scrolled off the viewport may be recycled and used with items scrolling into the viewport. When this occurs, the composition count needs to be reset. Otherwise, it may look like the item scrolling into the viewport has recomposed more than it actually has.

* Fix code style