Fixing compile error in DeviceCompatibilityModeTestJavaSnippets (#575)
Happened to notice a typo in the file, but it still didn't compile when I fixed the
typo, so I made a second fix.

The typo was a missing semicolon at the "import AppCompatActivity" line. But after
I fixed that, I got an error complaining about `assertFalse(isLetterboxed(activity))`
because `isLetterboxed` is expecting an AppCompatActivity but `activity`
is a `MainActivity`. Changing `isLetterboxed` fixed the compile error and also
meant we don't import AppCompatActivity so the original typo goes away. [shrug emoji]

Is this fix okay?