Enable spotless on all modules (#587)
* Enable spotless on all modules

The `./gradlew spotlessApply` command was failing in the
`:watchfacepush:validator` and `:identity:credentialmanager` modules
due to incorrect path calculations during the format and apply steps.
This was caused by ambiguous source directory configurations in their
`build.gradle.kts` files, which led to errors like `NoSuchFileException` and `FileAlreadyExistsException`.

This commit resolves these issues by explicitly defining the
`sourceSets` in the build scripts for the affected modules.
This removes the ambiguity and ensures that Gradle and the
Spotless plugin can reliably locate the source files.

As a result of this fix, the `apply_spotless.yml` GitHub workflow has
been simplified to run a single, project-wide `spotlessApply` command.

The `.gitignore` has also been updated to correctly ignore all `build` directories.

* Apply Spotless

---------

Co-authored-by: cartland <846051+cartland@users.noreply.github.com>