Disabling some "unused" warnings
I'd submitted code that's never actually used in the android/snippets apps. Ideally we'd include it in an app, but in the meantime, disabling the "unused" warnings.

Also removed one line (outside of a published-on-DAC code block) that turned out not to be needed.

No changes to the code that's included on DAC, and the code still compiles.